Adding HotFix for 'not found' error #11
This is referring to the issue reported here #11.
I encountered this error myself and made this quick fix to get the program to finish.
@rwick Might be good to incorporate a fix since it currently dies rather suddenly.

I haven't figured out yet why the reads (I've got 587 out of 2.4 mil) are not being classified in the first place (the couple that I blasted seem like real data). If I figure it out I'll report in a separate "non classified reads" issue about this.